+++ casual/casual-api/src/main/java/se/laz/casual/api/util/work/RepeatUntilSuccessTaskWork.java
@@ -4,7 +4,7 @@
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
 
-package se.laz.casual.api.util;
+package se.laz.casual.api.util.work;
 
 import javax.resource.spi.work.Work;
 import javax.resource.spi.work.WorkException;
