+++ casual/casual-jca/src/main/java/se/laz/casual/jca/work/StartInboundServerListener.java
@@ -50,7 +50,6 @@ public class StartInboundServerListener implements WorkListener
         if(null != e.getException())
         {
             log.warning(() -> "workCompleted with exception: " + e.getException());
-            log.warning(() -> "xxx Inbound may not have started!!!");
         }
     }
 }
