+++ casual/casual-network/src/main/java/se/laz/casual/network/inbound/reverse/AutoConnect.java
@@ -1,6 +1,6 @@
 package se.laz.casual.network.inbound.reverse;
 
-import se.laz.casual.api.util.RepeatUntilSuccessTaskWork;
+import se.laz.casual.api.util.work.RepeatUntilSuccessTaskWork;
 import se.laz.casual.network.reverse.inbound.ReverseInboundConnectListener;
 import se.laz.casual.network.reverse.inbound.ReverseInboundListener;
 import se.laz.casual.network.reverse.inbound.ReverseInboundServer;
