+++ casual/casual-network/src/main/java/se/laz/casual/network/outbound/EventLoopFactory.java
@@ -8,7 +8,6 @@ package se.laz.casual.network.outbound;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.epoll.EpollEventLoopGroup;
 import io.netty.channel.nio.NioEventLoopGroup;
-import se.laz.casual.api.util.JEEConcurrencyFactory;
 import se.laz.casual.config.ConfigurationService;
 import se.laz.casual.config.Outbound;
